requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
cryptography
web3
